fix(chat): move the voice chat types inside the voice chat
